rm tracker fns from ingest_permissions.py and place in ingest_trackers.py
